requirements.txt
streamlit
pyalex
pandas
plotly
pyarrow
numpy
scipy
umap-learn
psycopg2-binary
metric_learn
scikit-learn
openpyxl
requests
python-dotenv
sentence-transformers
igraph
leidenalg
hdbscan
clickhouse-connect
neo4j
